requirements.txt
torch
torchvision
gradio
pillow
numpy
scikit-learn
matplotlib